+++ statina/API/external/api/api_v1/endpoints/login.py
@@ -4,9 +4,13 @@ from typing import Optional
 from fastapi import APIRouter, Depends
 from fastapi.responses import RedirectResponse
 from fastapi.security import OAuth2PasswordRequestForm
+from sendmail_container import FormDataRequest
 
 from statina.API.external.api.deps import authenticate_user, create_access_token
-from statina.config import settings
+from statina.adapter import StatinaAdapter
+from statina.config import settings, get_nipt_adapter, email_settings
+from statina.crud import update
+from statina.crud.find import find
 from statina.models.database import User
 
 router = APIRouter()
@@ -28,6 +32,27 @@ def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()) -> 
         form_data=form_data,
         expires_delta=access_token_expires,
     )
+
+
+@router.put("/validate_user/{username}/{verification_hex}")
+async def validate_user(
+    username: str,
+    verification_hex: str,
+    adapter: StatinaAdapter = Depends(get_nipt_adapter),
+):
+    update_user: User = find.user(user_name=username, adapter=adapter)
+    if update_user.verification_hex == verification_hex:
+        update_user.role = "inactive"
+        update.update_user(adapter=adapter, user=update_user)
+        email_form = FormDataRequest(
+            sender_prefix=email_settings.sender_prefix,
+            request_uri=email_settings.mail_uri,
+            recipients=email_settings.admin_email,
+            mail_title="New user request",
+            mail_body=f"User {update_user.username} ({update_user.email}) requested new account <br>"
+            f'Follow <a href="{email_settings.website_uri}">link</a> to activate user',
+        )
+        email_form.submit()
 
 
 @router.get("/logout")
