+++ statina/models/database/user.py
@@ -9,6 +9,12 @@ class User(BaseModel):
     username: str
     email: EmailStr
     added: datetime
-    role: Literal["R", "RW", "inactive", "admin", "unconfirmed"]
+    role: Literal[
+        "unconfirmed",
+        "inactive",
+        "R",
+        "RW",
+        "admin",
+    ]
     hashed_password: str
     verification_hex: str = secrets.token_hex(64)
